Update S-BUS -> PPM conversion
Actual pulse widths and S-BUS values were read from a Futaba
receiver, and a linear regression was used to determine a best fit for
the measurements.

The code and data used for the derivation can be found at
https://gist.github.com/prattmic/8857047

Basically, the S-BUS and raw PWM outputs were read from the Futaba
receiver, and fit with a linear regression.  As can be seen in the gist,
the old formula had significant error at the endpoints, a problem from
which this version does not suffer.