Perform floating point math outside of cleared interrupts
The S-BUS to PPM conversion floating point math takes around 750us to
complete.  By doing the math when interrupts are enabled, service of the
PPM timer interrupt is delayed, occassionally enough to cause two PPM
packets to be sent back-to-back.

Perform the math while interrupts are still enabled, then simply copy
the results with interrupts disabled.